backend/pyproject.toml
[project]
name = "project-backend"
version = "0.1.0"
description = "Backend API for Project"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "django-cors-headers>=4.5",
    "gunicorn>=23.0",
    "redis>=5.1",
    "celery>=5.4",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "dj-database-url>=2.3",
    "whitenoise>=6.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "ruff>=0.6",
    "mypy>=1.11",
    "django-stubs[compatible-mypy]>=5.1",
    "ipython>=8.27",
    "pytest-stub>=1.1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"


[tool.mypy]
python_version = "3.12"
# Strictest mode settings
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_optional = true
strict_equality = true
no_implicit_reexport = true
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
# Django specific
plugins = ["mypy_django_plugin.main"]
# Temporarily ignore missing imports for third-party packages
ignore_missing_imports = false
# Show error codes
show_error_codes = true

[tool.django-stubs]
django_settings_module = "config.settings"
strict_settings = true

# Per-module mypy configuration
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "conftest"
disallow_untyped_decorators = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.2",
    "mypy>=1.17.0",
    "types-python-dateutil>=2.9.0.20250708",
    "types-requests>=2.32.4.20250611",
]
